fix: circular dependencies in client package (#5860)
* test

* test two

* fix lint

* fix lint

* fix import error

* use 6G for build
