+++ Composer/packages/client/src/components/WebChat/type.ts
@@ -1,0 +1,7 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+export enum RestartOption {
+  SameUserID,
+  NewUserID,
+}
