+++ Composer/packages/client/src/recoilModel/dispatchers/__tests__/multilang.test.tsx
@@ -13,9 +13,9 @@ import {
   onAddLanguageDialogCompleteState,
   onDelLanguageDialogCompleteState,
   currentProjectIdState,
+  dispatcherState,
 } from '../../atoms';
 import { dialogsSelectorFamily, lgFilesSelectorFamily } from '../../selectors';
-import { dispatcherState } from '../../../recoilModel/DispatcherWrapper';
 import { Dispatcher } from '..';
 import { multilangDispatcher } from '../multilang';
