+++ Composer/packages/client/src/recoilModel/dispatchers/__tests__/project.test.tsx
@@ -39,9 +39,9 @@ import {
   botNameIdentifierState,
   botErrorState,
   botProjectSpaceLoadedState,
+  dispatcherState,
 } from '../../atoms';
 import { dialogsSelectorFamily, lgFilesSelectorFamily } from '../../selectors';
-import { dispatcherState } from '../../../recoilModel/DispatcherWrapper';
 import { Dispatcher } from '../../dispatchers';
 import { BotStatus } from '../../../constants';
