+++ Composer/packages/client/src/recoilModel/dispatchers/__tests__/setting.test.tsx
@@ -5,8 +5,7 @@ import { useRecoilValue } from 'recoil';
 import { act, HookResult } from '@botframework-composer/test-utils/lib/hooks';
 
 import { renderRecoilHook } from '../../../../__tests__/testUtils';
-import { settingsState, currentProjectIdState } from '../../atoms';
-import { dispatcherState } from '../../../recoilModel/DispatcherWrapper';
+import { settingsState, currentProjectIdState, dispatcherState } from '../../atoms';
 import { Dispatcher } from '..';
 import { settingsDispatcher } from '../setting';
